backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bootstrap4==0.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
config==0.5.1
confusable-homoglyphs==3.2.0
dj-database-url==2.0.0
Django==4.2.2
django-bootstrap4==23.1
django-database-url==1.0.3
django-filter==23.2
django-paypal==2.0
django-phonenumber-field==7.1.0
django-registration==3.3
django-taggit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
mysqlclient==2.1.1
phonenumbers==8.13.15
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
